% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cmdline.R
\name{install_cmdline}
\alias{install_cmdline}
\title{install_cmdline}
\usage{
install_cmdline(profile_path = "~/.bash_profile")
}
\arguments{
\item{profile_path}{path to profile (default: ~/.bash_profile)}
}
\value{

}
\description{
install_cmdline
}
